refactor(TagCreationForm.svelte): update import statements for tagLocalStorageManager and doesTagExist to reflect new file structure
refactor(TagCreationForm.svelte): replace usage of addTagToLocalStorage with tagLocalStorageManager.addItem to improve code organization and readability

refactor(TagEditForm.svelte): update import statements for tagLocalStorageManager and doesTagExist to reflect new file structure

refactor(TagEditForm.svelte): replace usage of updateTagInLocalStorage with tagLocalStorageManager.updateItem to improve code organization and readability

refactor(TagItem.svelte): update import statements for deleteItemFromDatabaseRequest and tagLocalStorageManager to reflect new file structure

refactor(TagItem.svelte): replace usage of deleteTagFromLocalStorage with tagLocalStorageManager.deleteItem to improve code organization and readability

refactor(TagList.svelte): update import statements for deleteItemFromDatabaseRequest and tagLocalStorageManager to reflect new file structure

refactor(TagList.svelte): replace usage of deleteAllTagsFromLocalStorage with tagLocalStorageManager.deleteAllItems to improve code organization and readability

refactor(darkModePreferenceStore.ts): remove import statement for LocalStorageKeys as it is no longer used

refactor(tagStore.ts): update import statement for get to reflect new file structure

refactor(tagStore.ts): moved doesTagExist function to tagStore.ts to improve code organization

refactor(databaseUtils.ts): create a new file for deleteItemFromDatabaseRequest function to improve code organization

feat(localStorageManager.ts): create storage manager to handle operations on local storage, including adding, updating, deleting, and retrieving items

refactor(tagLocalStorageMethods.ts): deleted file as it is no longer needed

chore(page.server.ts): add error handling for undefined or null tagId in update and delete operations

+++ src/routes/(app)/dashboard/tags/+page.server.ts
@@ -7,6 +7,11 @@ import type { Actions, PageServerLoad } from './$types';
 
 type TagOperation = 'create' | 'update' | 'delete' | 'deleteAll';
 
+interface TagData {
+    name: string;
+    user_id: string;
+}
+
 import { TagValidationSchema } from '$dashboardUtils/validation/tagValidationSchema';
 
 /**
@@ -50,8 +55,13 @@ const handleTagOperation = async (
     tagForm?: SuperValidated<typeof TagValidationSchema>,
     tagId?: string | null
 ) => {
+    // Check for undefined or null tagId for update and delete operations
+    if ((tagOperation === 'update' || tagOperation === 'delete') && !tagId) {
+        console.error('Tag ID is undefined or null');
+        return { status: 400, body: { message: 'Tag ID is required' } };
+    }
     // Prepare tag data if it's not a delete or deleteAll operation
-    const tagData =
+    const tagData: TagData | Record<string, never> =
         tagForm && tagOperation !== 'delete' && tagOperation !== 'deleteAll'
             ? { name: tagForm.data.name, user_id: session.user.id }
             : {};
@@ -116,7 +126,7 @@ export const actions: Actions = {
 
     deleteTag: async ({ request, locals: { getSession, supabase } }) => {
         const formData = new URLSearchParams(await request.text());
-        const tagId = formData.get('tagId');
+        const tagId = formData.get('itemId');
 
         const session = await getSession();
